requirements.txt
# Web Server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Camera
opencv-python==4.9.0.80

# Storage
pyyaml==6.0.1

# Utilities
pydantic==2.5.3
